feat: pss-imp-vis-03-response-event-presentation-001 - Nest response_event_presentation behind the Visualization root facade
