+++ pkg/services/factory_visualization/root_contract.go
@@ -507,9 +507,9 @@ func (s *Service) OpenPresentation(
 	var output Output
 	switch req.Mode {
 	case PresentationDeliveryBestEffort:
-		output = newBestEffortOutput(writer)
+		output = openBestEffortOutput(writer)
 	case PresentationDeliveryLossless:
-		output = newLosslessOutput(writer)
+		output = openLosslessOutput(writer)
 	default:
 		return OpenPresentationResult{}, &PresentationError{
 			Kind:    PresentationErrorInvalidInput,
@@ -617,7 +617,7 @@ func (s *Service) FinalizePresentation(
 		session.closed = true
 		return FinalizePresentationResult{}, err
 	}
-	if _, err := session.writer.Write(appendLine(req.Terminal.Payload)); err != nil {
+	if _, err := session.writer.Write(appendPresentationLine(req.Terminal.Payload)); err != nil {
 		session.finalized = true
 		session.closed = true
 		return FinalizePresentationResult{}, err
@@ -695,10 +695,3 @@ func requirePresentationContext(ctx context.Context) error {
 	return nil
 }
 
-func isPresentationClosedErr(err error) bool {
-	return errors.Is(err, errPresentationOutputClosed)
-}
-
-func isPresentationBackpressureErr(err error) bool {
-	return errors.Is(err, errPresentationBacklogFull)
-}
